Use modulus, fix types/printf, add READMEs
Replace integer division with remainder where intended (homework 3zad/4zad), fix type and printf mismatches across lesson and homework examples (convert several int variables to float, cast division to float, and update printf format specifiers), and add README files for both Lesson and Homework directories documenting tasks and compilation notes. Key changes: homework 1zad printf specifiers changed, homework 3zad/4zad use a % b, lesson 3zad/4zad/5zad adjust variable types and prints, and two README.md files were added to describe the exercises.

+++ Year 1/Semester 1/Introduction to Programming/IfElse/IfElse-27.10.2023_Lesson/5zad/5zad.cpp
@@ -3,30 +3,30 @@
 
 int main()
 {
-    int y, x;
+    float y, x;
 
     printf("x = ");
     scanf_s("%d", &x);
 
     if (x <= 0) {
         y = 2;
-        printf("y = %d", y);
+        printf("y = %f", y);
     }
     else if (x > 0 && x < 1) {
         y = x + 2;
-        printf("y = %d", y);
+        printf("y = %f", y);
     }
     else if (x >= 1 && x <= 2) {
         y = 3;
-        printf("y = %d", y);
+        printf("y = %f", y);
     }
     else if (x > 2 && x < 3) {
         y = 5 - x;
-        printf("y = %d", y);
+        printf("y = %f", y);
     }
     else if (x >= 3) {
         y = 2;
-        printf("y = %d", y);
+        printf("y = %f", y);
     }
 
     return 0;
